Improve output paths

+++ .nuke/BuildProject.cs
@@ -6,23 +6,20 @@ using Nuke.Common.IO;
 using Nuke.Common.Tools.DotNet;
 using Serilog;
 
+namespace _build;
+
 // ReSharper disable UnusedMember.Local
 // ReSharper disable AllUnderscoreLocalParameterName
 class BuildProject : NukeBuild
 {
-
     public static int Main() => Execute<BuildProject>();
 
     #region Parameters
 
     [Parameter("Verbosity level of the build output - Default is 'Minimal'")]
     readonly DotNetVerbosity DotNetVerbosityLevel = DotNetVerbosity.quiet;
-    
-    [Parameter("Skips the build step if specified.")]
-    readonly bool NoBuild;
 
-    [Parameter("Paths to the test project(s) to run. You can provide multiple paths as separate arguments.")]
-    readonly AbsolutePath[] TestProjectPaths;
+    [Parameter("Skips the build step if specified.")] readonly bool NoBuild;
 
     #endregion
 
@@ -38,76 +35,64 @@ class BuildProject : NukeBuild
                 Log.Information("Cleaning directory {Directory}", directory);
                 directory.DeleteDirectory();
             }
-            
-            var buildDirectory = RootDirectory / "build";
-            Log.Information("Cleaning directory {BuildDirectory}", buildDirectory);
-            buildDirectory.DeleteDirectory();
 
-            var publishDirectory = RootDirectory / "pub";
-            Log.Information("Cleaning directory {PublishDirectory}", publishDirectory);
-            publishDirectory.DeleteDirectory();
+            var pubCliDirectory = RootDirectory / "pubcli";
+            Log.Information("Cleaning directory {PublishDirectory}", pubCliDirectory);
+            pubCliDirectory.DeleteDirectory();
+
+            var pubWebDirectory = RootDirectory / "pubweb";
+            Log.Information("Cleaning directory {PublishDirectory}", pubWebDirectory);
+            pubWebDirectory.DeleteDirectory();
         });
 
-    Target Deploy => _ => _
-        .DependsOn(Test, Publish)
+    Target DeployCli => _ => _
+        .DependsOn(Test, PublishCli)
         .Executes(() =>
         {
             var appDataPath = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData);
             var targetDirectory = Path.Combine(appDataPath, "utils");
-            if (!Directory.Exists(targetDirectory))
-            {
-                Directory.CreateDirectory(targetDirectory);
-            }
+            if (!Directory.Exists(targetDirectory)) Directory.CreateDirectory(targetDirectory);
 
             const string fileName = "reddit.exe";
             var targetFile = Path.Combine(targetDirectory, fileName);
-            if (File.Exists(targetFile))
-            {
-                File.Delete(targetFile);
-            }
-            
-            var sourceFile = RootDirectory / "pub" / "PersonalFinance.CLI.exe";
+            if (File.Exists(targetFile)) File.Delete(targetFile);
+
+            var sourceFile = RootDirectory / "pubcli" / "app.exe";
             File.Copy(sourceFile, targetFile);
             Log.Information("Deployed {TargetFile} to {TargetDirectory}", fileName, targetDirectory);
         });
 
-    Target Publish => _ => _
+    Target PublishCli => _ => _
         .DependsOn(DeepClean)
-        .Before(Deploy)
+        .Before(DeployCli)
         .Executes(() =>
         {
             DotNetTasks.DotNetPublish(s => s
-                .SetProject(RootDirectory / "PersonalFinance.sln")
+                .SetProject(RootDirectory / "src" / "app" / "app.csproj")
                 .SetConfiguration(Configuration.Release)
                 .SetVerbosity(DotNetVerbosityLevel)
-                .SetOutput(RootDirectory / "pub"));
+                .SetOutput("pubcli"));
+        });
+
+    Target PublishWeb => _ => _
+        .DependsOn(DeepClean)
+        .Before(DeployCli)
+        .Executes(() =>
+        {
+            DotNetTasks.DotNetPublish(s => s
+                .SetProject(RootDirectory / "src" / "ui" / "ui.csproj")
+                .SetConfiguration(Configuration.Release)
+                .SetVerbosity(DotNetVerbosityLevel)
+                .SetOutput("pubweb"));
         });
 
     Target Test => _ => _
         .Executes(() =>
         {
-            GetTestProjects().AsParallel().ForAll(project =>
-            {
-                DotNetTasks.DotNetTest(s => s
-                    .SetProjectFile(project)
-                    .SetVerbosity(DotNetVerbosityLevel));
-            });
+            DotNetTasks.DotNetTest(s => s
+                .SetProjectFile(RootDirectory / "reddit.sln")
+                .SetVerbosity(DotNetVerbosityLevel));
         });
 
     #endregion
-
-    #region Helper Methods
-
-    AbsolutePath[] GetTestProjects() =>
-        (TestProjectPaths ?? []).Length > 0
-            ? TestProjectPaths
-            :
-            [
-                RootDirectory / "tests" / "PersonalFinance.Application.Tests" / "PersonalFinance.Application.Tests.csproj",
-                RootDirectory / "tests" / "PersonalFinance.Domain.Tests" / "PersonalFinance.Domain.Tests.csproj",
-                RootDirectory / "tests" / "PersonalFinance.Infrastructure.Tests" / "PersonalFinance.Infrastructure.Tests.csproj"
-            ];
-
-    #endregion
-
 }
